fix: Revise publication logic for better data handling & querying
Body: The changes in this commit address two main points: revising the method of querying list of chapters and enhancing the logic of deleting and storing chapter related data.

Firstly, the querying for chapters was previously done by ID. This method was dependent on receiving a list of IDs, which is not very flexible or extensible. Now, the chapter repository query method retrieves chapters using the memberID and parentChapterId, improving the modularity of the function.

Secondly, multiple changes were made to improve handling of chapter related data deletion and storage. Instead of a fragmented approach, where different parts of data associated with a chapter were deleted in several functions, a comprehensive function was introduced to handle the deletion of all related data. This makes the logic more intuitive, maintainable, and less error-prone.

Additionally, initial allocation of BookChapter and BookContent data containers was moved to the constructor, ensuring that it is never null and simplifying the processing logic.

Thirdly, the PublicationController now consumes MULTIPART_FORM_DATA type value providing more efficient data processing with form data.

Finally, the Publication and its associated Book are being saved to the repository immediately after their creation. This ensures that these entities are persisted before Chapter related data is deleted, mitigating potential data inconsistencies.